Finish definition of locally ringed space
Add the skeleton of what it means to be a
sheaf, enough to finish the (first iteration
of the) definition of locally ringed space.

The notion of a sheaf is **unbundled** as
of now, with an eye towards situations
like cokernels of sheaves and non-noetherian
adic spaces.

Also added a definition for the restriction
maps in presheaf.lean